fix(mqtt): consolidate Mqtt/ into MQTT/ to fix case-collision build break
PR #4138 added src/components/Mqtt/ while the established directory is
src/components/MQTT/. On case-sensitive Linux both resolve, so Vitest/CI
Quick Tests passed — but on case-insensitive macOS/Windows filesystems
TypeScript sees the same directory referenced under two casings and fails
the Desktop CI frontend build with TS1261. Move the six packet-monitor
files into MQTT/ and point App.tsx's import at the canonical casing.
